Add a VS2015 solution
Created a VS2015 solution. Had to create solutions for the code and
dependencies. Also includes a fix to find the right OpenGL header, as
well as some fixes for VS-specific compile errors.

Output is placed in the `build` folder in the project root. To use the
local debugger make sure that the cake3d project is the startup project
and that the working directory is set to $(SolutionDir)

This has been tested with VS2015 Community Edition.